feat(ШАГ 6): Рефакторинг ParallelBranchTab.tsx - разбиение на компоненты
 СОЗДАНО:
- useParallelBranchLogic.ts (35 строк) - хук для фильтрации вопросов
- BranchBasicSettings.tsx (40 строк) - основные настройки ветки
- BranchRepeatSettings.tsx (71 строк) - настройки повторений и режима отображения
- BranchCountSettings.tsx (43 строк) - настройки поля количества
- BranchQuestionsList.tsx (104 строки) - управление списком вопросов с drag&drop

 ИЗМЕНЕНО:
- ParallelBranchTab.tsx: 285  75 строк (-74% кода!)
- Убраны все inline-компоненты и дублированная логика
- Применены принципы Single Responsibility и композиции
- Вынесена логика drag&drop в отдельный компонент

 РЕЗУЛЬТАТ:
- Сокращение на 210 строк кода в основном файле
- 5 переиспользуемых модульных компонентов
- Функциональность полностью сохранена
- TypeScript компиляция без ошибок
