+++ src/components/QuestionForm/components/index.ts
@@ -1,0 +1,4 @@
+export { BranchBasicSettings } from './BranchBasicSettings';
+export { BranchRepeatSettings } from './BranchRepeatSettings';
+export { BranchCountSettings } from './BranchCountSettings';
+export { BranchQuestionsList } from './BranchQuestionsList';
